chore: setup changes for notebooks (#21)
Introduces 2 changes:

* `setup.sh` will optionally read 2 environment variables if set.
${BIOEMU_ENV_NAME}, which is the name conda environment bioemu will be
installed in, and ${UPDATE_ENV}, to force update of an existing
environment. The latter argument is introduced because in Colab
notebooks, only the `base` environment is available to the user via
condacolab, and switching to a different environment (jupyter kernel)
from within the notebook is not possible.
* `environment.yml` now lists Python>3.10 as a dep instead of pinning it
down to 3.12. It's possible that changing the channel priority
restriction argument might also fix the resolution error.